refactor(storybook): one Storybook project for the whole workspace
Replaces the two separate Storybooks with a single `storybook` Nx project
publishing to one Chromatic project.

The viewer owned a Storybook and the design system was about to get a second
one, which would have meant two Chromatic projects for one product. It also hid
the interaction between them: the platform app loads `globals.css` and then the
viewer's stylesheet as a lazy chunk, and that ordering is exactly what let the
viewer overwrite the host's radius scale after hydration. The preview now loads
both, in that order, so a regression of that kind is a diff rather than a
production surprise.

Stories stay co-located with the code they document. The new project owns only
configuration and globs both source trees. Titles are namespaced `Foundations/*`,
`Components/*` and `Viewer/*`; the viewer's two `Components/*` stories moved
under `Viewer/*`, where they no longer collide with the shared UI.

Stories render light above dark in one snapshot, since half the defects worth
catching only appear in one theme. The viewer's opt out: they scope their own
theme and each mounts a WebGL canvas.

Each project still type-checks its own stories with its own compiler settings.
Pulling the viewer's stories into the design system's tsconfig dragged
`@vctrl/core` in under the wrong `lib` and failed on `new Error(msg, { cause })`.

Fixes six type errors in the viewer's overlay stories, which had never been
checked: `build-storybook` depended on `typecheck`, and that target uses
`tsconfig.lib.json`, which excludes stories. They passed a `hasContent` prop
that `Overlay` does not accept, and never passed `loadingState`, which it
requires - so `Loading` and `Loaded` rendered the same undefined state and their
snapshots meant nothing.

Also ignores `**/storybook-static/**` in ESLint; the directory is gitignored but
lint globs still reached the built bundle.
