+++ .github/workflows/cd-chromatic.yaml
@@ -1,14 +1,18 @@
-name: CD - Publish Viewer to Chromatic
+name: CD - Publish Storybook to Chromatic
 
 on:
   push:
     branches:
       - main
     paths:
+      - 'storybook/**'
+      - 'shared/components/**'
       - 'packages/viewer/**'
-      - '.github/workflows/cd-chromatic-viewer.yaml'
+      - '.github/workflows/cd-chromatic.yaml'
   pull_request:
     paths:
+      - 'storybook/**'
+      - 'shared/components/**'
       - 'packages/viewer/**'
 
 concurrency:
@@ -21,9 +25,14 @@ permissions:
 
 jobs:
   chromatic:
-    name: Publish to Chromatic
+    name: Publish Storybook to Chromatic
     runs-on: ubuntu-22.04
     environment: chromatic-publishing
+
+    # One Chromatic project for the whole workspace. The upload is skipped when
+    # no token is configured, so the build and typecheck still run either way.
+    env:
+      CHROMATIC_TOKEN: ${{ secrets.CHROMATIC_PROJECT_TOKEN_1 }}
 
     steps:
       - name: ⬇️ Checkout repo
@@ -44,13 +53,19 @@ jobs:
         run: pnpm install --frozen-lockfile
 
       - name: 🔨 Build Storybook
-        run: npx nx build-storybook vctrl/viewer
+        run: npx nx build-storybook storybook
+
+      - name: ⏭️ Chromatic token not configured
+        if: env.CHROMATIC_TOKEN == ''
+        run: |
+          echo "::notice::CHROMATIC_PROJECT_TOKEN_1 is not set. Storybook was built but not published."
 
       - name: ⚡ Publish to Chromatic
+        if: env.CHROMATIC_TOKEN != ''
         uses: chromaui/action@v18
         with:
-          projectToken: ${{ secrets.CHROMATIC_PROJECT_TOKEN_1 }}
-          workingDir: packages/viewer
+          projectToken: ${{ env.CHROMATIC_TOKEN }}
+          workingDir: storybook
           autoAcceptChanges: 'main'
           exitZeroOnChanges: ${{ github.event_name == 'pull_request' }}
           exitOnceUploaded: true
