+++ packages/viewer/package.json
@@ -43,8 +43,6 @@
 		"./css": "./style.css"
 	},
 	"scripts": {
-		"build-storybook": "storybook build",
-		"chromatic": "npx chromatic --project-token=chpt_b3f4d26ac3d3190",
 		"type-check": "tsc --noEmit"
 	},
 	"dependencies": {
