+++ shared/components/project.json
@@ -62,35 +62,6 @@
 					"mode": "production"
 				}
 			}
-		},
-		"build-storybook": {
-			"executor": "nx:run-commands",
-			"dependsOn": [
-				"typecheck",
-				"typecheck-storybook"
-			],
-			"outputs": [
-				"{projectRoot}/storybook-static"
-			],
-			"options": {
-				"command": "storybook build",
-				"cwd": "shared/components"
-			}
-		},
-		"storybook-dev": {
-			"executor": "nx:run-commands",
-			"continuous": true,
-			"options": {
-				"command": "storybook dev",
-				"cwd": "shared/components"
-			}
-		},
-		"typecheck-storybook": {
-			"executor": "nx:run-commands",
-			"outputs": [],
-			"options": {
-				"command": "tsc --noEmit -p shared/components/tsconfig.storybook.json"
-			}
 		}
 	}
 }
