--- shared/components/tsconfig.storybook.json
@@ -1,28 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"emitDecoratorMetadata": true,
-		"outDir": "",
-		"jsx": "react-jsx",
-		"moduleResolution": "bundler"
-	},
-	"files": [
-		"../../node_modules/@nx/react/typings/styled-jsx.d.ts",
-		"../../node_modules/@nx/react/typings/cssmodule.d.ts",
-		"../../node_modules/@nx/react/typings/image.d.ts"
-	],
-	"exclude": [
-		"src/**/*.spec.ts",
-		"src/**/*.test.ts",
-		"src/**/*.spec.tsx",
-		"src/**/*.test.tsx"
-	],
-	"include": [
-		"src/**/*.stories.ts",
-		"src/**/*.stories.jsx",
-		"src/**/*.stories.tsx",
-		"src/**/*.stories.mdx",
-		".storybook/*.ts",
-		".storybook/*.tsx"
-	]
-}
