+++ storybook/package.json
@@ -1,0 +1,10 @@
+{
+	"name": "@vectreal/storybook",
+	"version": "0.0.1",
+	"private": true,
+	"type": "module",
+	"scripts": {
+		"build-storybook": "storybook build",
+		"storybook": "storybook dev"
+	}
+}
