+++ storybook/project.json
@@ -1,0 +1,43 @@
+{
+	"name": "storybook",
+	"$schema": "../node_modules/nx/schemas/project-schema.json",
+	"sourceRoot": "storybook",
+	"projectType": "application",
+	"tags": [],
+	"targets": {
+		"typecheck": {
+			"executor": "nx:run-commands",
+			"outputs": [],
+			"options": {
+				"command": "tsc --noEmit -p storybook/tsconfig.json"
+			}
+		},
+		"build-storybook": {
+			"executor": "nx:run-commands",
+			"dependsOn": [
+				"typecheck",
+				{
+					"projects": [
+						"vctrl/viewer"
+					],
+					"target": "typecheck-storybook"
+				}
+			],
+			"outputs": [
+				"{projectRoot}/storybook-static"
+			],
+			"options": {
+				"command": "storybook build",
+				"cwd": "storybook"
+			}
+		},
+		"storybook-dev": {
+			"executor": "nx:run-commands",
+			"continuous": true,
+			"options": {
+				"command": "storybook dev",
+				"cwd": "storybook"
+			}
+		}
+	}
+}
